requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0
django-crispy-forms==1.13.0
django-filter==21.1
django-related-admin==1.0.1
djangorestframework==3.13.1
flatdict==4.0.1
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.9
